This commit adds ancilliary functionality necessary to get the run-time coupling development off the ground. These changes create and link a new library (libRTC.a), which will contain either the associated coupling code, or fortran stubs that disable the coupling functionality. The Install script now supports a command-line argument ("--RTC") that allows inclusion of this experimental feature. And ESP-r's read/write routines now recognize a *RTC-XXXX keyword in the .cfg file, where XXXX denotes the type of coupling to be invoked. At present, one keyword ("*RTC-demo") is supported, which will eventually cause bps to couple with one of the simple win32 programs in the Win32_apps directory.
This work is building up to a COM-interface coupling between 
MinGW-compiled versions of ESP-r and other Win32-based software --- 
namely TRNSYS. But the basic structure proposed in this commit should 
provide a rational platform for other couplings as well, perhaps via 
fork-exec and unix pipes.

Note that these changes do not actually contain any run-time coupling 
code, which admittedly is the daunting part of this project. Forward 
ho!




Former-commit-id: 07825ec5a143c3b6dda30abf4e5994c9988113c7