This commit adds flow network control w/o plant based on sensed R.H. of a zone air point. First tests imply that all controllers that can be used for temperature based flow network control can also be used for R.H. (setpoint dimension may be critical for some controllers).
To use: Set up temperature based controller for zone desired. Then change <zone>  0   0    0  in .ctl to <zone>  0   0   -66. Also, change controller type from "1" to "2". The available type "11" is only used in context with a flow network *and* a plant, seemingly.

Compile and run on Cygwin O.K.
Simple model results O.K.
Further testing r'qd.

It is not possible to enter / edit this control with Project Manager.



Former-commit-id: afe4de151ab7e8aa86cc5a5352404b9d09299af0