This commit adresses issues in module dfs and in cfd - afn conflation.
- Included bug-fix done by Aizaz (editing of afn - dfs conflation)
- changed minimum number of necessary iterations from 20 to 5
- changed format of write to monitor file to single-line
- introduced logic to turn off interruption of multiple time-step simulations when using afn conflation. This - of course - should be used with care. Following changes were implemented:
   a) added parameter NOSTOP to common /MFS/
   b) moved parameter MITMFS to common /MFS/ (was a simple "parameter()", befor
   c) added read from .dfs file and (default) write to .dfs file of these parameters
NOSTOP, when 0, does nothing. When .ne. 0, the simulation is not interrupted. If convergence is not reached, this is ignored. MITMFS sets the maximum number of loops in which conflation agreement between dfs and mfs is tried to reach. Was 20 as default (defaults to this value, still).

Compile and run on cygwin (gcc 3.4.4) O.K.
Tests: Read file without new parameters O.K., write new parameters O.K., simple model (Nalin) runs through O.K.

A more usable text output (less numbers, more information) to the monitor file would be good for this feature, to be able to "recap" what the convergence behaviour was after a weekend-run, e.g. ...




Former-commit-id: 721a08dcba66a5a2980645d36ae074b3ee90260f