This commit wraps up work on new cfd option parameter NOSTOP and "editability" of MITMFS.
Added parameter edit to project manager.

No change to / impact on simulation results.

Compile and run on Cygwin O.K.

This commit together with commits 4242 and 4251 could be included in development_branch, from my point of view. As the features are not used in any tester and do not have any impact on simulation results, further "validation" testing is not possible, though.



Former-commit-id: cd623f7ff3f7bf3ee0f49044537a044f4b537e91