This commit more or less wraps up the new feature "multi timestep cfd - bsim conflation" with "NOSTOP" option. Writing to the monitor file was reduced (the actual output can still be improved in regard to information given about a run). Also, if NOSTOP is active, cfd results "surface convective coefficients" and "heat flux to surfaces" are passed to xml output (here, the call of lnblnk should be replaced by the recently introduced string length common variables).
Compile and run on OSX O.K.

Test model (Nalin) runs well, results and monitor file output as expected. 

These changes do not have any impact on existing models / simulations, as the new feature is not used therein and the "old" behaviour is as it was (save some write(6,*) statements which were commented out).




Former-commit-id: fc6b1142962b6c2f779a5934fabb5ba58f23ab10