This commit changes two variable names in e2r:
NRIF to NSCENE, as this is what the variable actually counts: The number of scene descriptions in the .rcf file. Each scene description can hold more than one .rif file, e.g. for blind control via .tmc. Many places in the code to date seem to allow for one additional .rif, only, however.

IRIFFOC to ISCENEFOC, as again this is what the index does: sets focus on current scene (which in many places is the same as current .rif, though - very confusing code). The whole logic seems to be restricted to looking at one zone of the model only (on dimensional name arrays etc.).

Compile and run on Cygwin gcc 4.3.2 + gfortran O.K.

Simple model browsing etc. no visible difference to before (which does not mean that it really works).



Former-commit-id: 8f770829e15900036430f6b54fc94f0efe460084