Update to miscel component database handling:
- shift dmcmpdbfl from common deflt3 and put into esprdbfile.h.
- change dmcmpdbfl from 72 characters to 144 char.
- change mcmpdbfl in esprdbfile.h to 144 char.
- search for places where mcmpdbfl is edited and update.
- update file browser to initially point to standard db folder.
- add ipathmsc to esprdbfile.h and update file I/O to reflect this
- copy optical db file browsing subroutine and adapt to work 
  with miscel components database.
- code tidy to remove unused variables and reduce syntax 
  reports

Users will now have access to the usual database copy
and browse factilites but not a create new.  The editing of the
miscel comp database is still via a text editor.

Because the miscel comp database (standard or local) is known
and managed, the scan of this database no longer needs to
ask the user for confirmation of the db file name.

This commit to allow syntax check on Linux machine and
compiling on other versions of compiler.

Testing
- Compile OSX GTK gcc 4.3 with databases and exemplars.
- Interactive tests - GTK browse works ok.  Selecting a component 
  type and setting up an active material for the model seems to 
  work as expected (types are listed and normal selection of 
  zone and surface).  Model cfg file recognises local vs
  standard database locations!




Former-commit-id: 69a7923a5ac414ed80d305b793fc41e9cc29a7ba