This commit addresses casual gains due to occupancy.
If one chooses "m2/p" as unit (ICGT(IDTY,I).eq.-1), the conversion factor from area to gain (W) is hard-coded to 95.0 W/P. This value is changed to 70 W/P to be conform to the (too low) value according to SIA load schedules.

Preferably, a variable "met" or "W/P" would be added to the .opr file, of course.

Also, a bounds error when extending the number of day types is addressed.

Compile and run on Cygwin gcc 4.3.2 + gfortran O.K.

Test runs (and interactive addition of day types) look good.



Former-commit-id: 32796b1876ed1a95d32994e79b78875ee34bda90