This commit is a revert to undo any modifications to the CETC Battery model located at src/esrupfs/CETC_BATTERY.F and any files it affects (see below).
To correct a convention issue regarding the sign of current 
and power reported to XML outputs in a battery discharge mode, a slight 
modification to the battery model has been made. 

Following the completion of quality assurance checklist, the 
modifications will be sent to the development branch for the latest 
release.




Former-commit-id: 6c227ea811d748927721963b8cad7e9c40527a73