This commit confirms the completion of quality assurance testingon the CETC battery model located at src/esrupfs/CETC_BATTERY.F following the current and power modification in a charge mode.
Summary of changes:

By convention, a demand imposed on the battery is assigned positive value if the battery is 
to discharge and a negative value if the battery is to charge. While the power value 
returned from the battery model reflected this convention, the XML and .csv outputting did 
not i.e. the power and current were reported as positive values regardless of if the battery 
was charging or discharging in XML outputting.

Consequently a minor modification was implemented in the subroutine to follow this 
convention in XML outputting. I.e. the following variables now report negative in XML 
outputting in a charge scenario:

-batCurFin, the battery current over the timestep
-pBatSystem, actual load on the battery system (part of the load that can be met)
-batPowBal, Balance of power supplied to/requested from battery and actual power consumption

Quality assurance has been completed by running tester.pl with the changes against the 
latest development branch with the battery tester cases confirming the change in variable 
sign.





Former-commit-id: ab2d338e80e5d81be429a71c42631fe1130a3da0