Modified mscomp.db1 to remove any lithium-ion power only component parameters.
In order to correct a convention oversight for battery current and power (see my previous 
commits) the component parameters in /src/databases/mscomp.db1 has to be the same as my 
original checkout.

I thought this was the case however upon comparing this database with the original checkout, 
it contained parameters for my new lithium-ion work. This commit undoes those database 
changes and reverts it to the current HEAD of the development branch.




Former-commit-id: c82dc030d878416364143f4a4b11a83327611b0e