This commit comments out the read_in_cfc_file() in edatabase.F added for QA reporting. For some reason, the code does not work on Cygwin (OS X was fine).
Looking at the code in read_in_cfc_file leads to questions in regard to the usage of the /G6/ commons in context with the >number of surfaces per zone<. The variables in this common seem to be defined for >number of total surfaces of model> or the like. This could explain the many messages received "not sure if ',SSNAME(I),' is a complex fenestration constr.' in models where the surfaces named really were no CFC and also shouldn't be. Possibly got past debugging due to effect only occurs in more-than-one-zone models where the different zones also have a different distribution of the CFCs across the zone surfaces ...

Compile and run on Cygwin O.K.



Former-commit-id: 39d0f4dadf833fd02542350caa462302eef310c8