Editorial changes in e2rform.F, extension of logic in several places to include CFCs in edcasctl.F.
The warning message "No TMC file for zone xx, blind control tracking not possible" should now not appear when using CFCs instead of TMCs. The TMC file would be rescanned, the CFC file is not. Not clear if this would be necessary here. Probably not.

Compile and run on Cygwin O.K.

Load model and enter cgc editing O.K.



Former-commit-id: 1fc267c9a010b62fc75d1d2f4792243ecbe8b27c