the component with code 690, three way diverger with control for water/steam, has been added to routines and database. The component is identical to component 650, but the fraction of flow is not fixed but controlled.
the component with code 1280, 3-node water storage stratified tank with 2HX, has been added to routines and database.
The component is identical to component 1260, but a 3rd node is added for the control of the water temperature at mid tank height and it allows inserting PCM modules.

the component with code 1290, 3-node water storage stratified tank with 2HX, has been added to routines and database.
The component is identical to component 1270, but a 4rd node is added for the control of the water temperature at mid tank height and it allows inserting PCM modules.


Former-commit-id: 831b80a07b1e6be51297d19201757490ae6a2c01