Added a test model related to retail buildings and which can be used to explore UK NCM performance issues.
This model approximates a retail box shed supermarket in the
UK with 2008 standard constructions.  There is diversity in
the internal gains. The public spaces are represented as a
foyer (with air mixing to the outside), main shop (with
internal mass of shelfs included), refrigerated section within
which is an abstract low temperature chilled cabinet and a
standard chilled cabinet (air is mixed with the cabinets to
approximate people accessing the cabinets).

There is a cafe on the right of the foyer and a newstand
and office on the left. The upper level includes an
office and a staff room with a passage connecting them.

The service areas at the back of the building include a
double height shipping and receiving area and storage
space for goods in transet and a small cold storage room.

Because of the large number of people that might be in the
building make-up air (6l/s/p) is passed through a tempered mixing
box (16C). The make-up air rate is reduced overnight. 

There are two models - a standard model and a variant which
includes a UK NCM description so that complience assessments
can be carried out.

The commit allows this model to be tested on other platforms
as well as for possible later inclusion in standard code testing
process.




Former-commit-id: 4dfd3a7cfc3f837f7b6b65223101fedefe963fac