Summary
A Scottish residential test model has been added to the
  validation/benchmark/DSM folder.
 
This model represents one half of a four unit dwelling (a 
lower 1BR flat and an upper 1BR flat with access stair).  
Its composition is representative of timber-kit energy 
efficient public housing in the late 1990s in the North of 
Scotland.  Walls and ceilings are well insulated and some
effort has been put into reducing infiltration. Windows
are standard double glazed units with wood frames.

Each room is explicitly represented as is the roof space 
and the crawl space. This allows transient occupancy patterns
to be studied as well as the impact on different infiltration
regimes.

This is an initial version of the model with the zone geometry
and composition specified. Room occupancy and environmental
controls have not yet been defined.  




Former-commit-id: 313087144ba5b248f6baaa2f21dc396012440539