Summary
Added another testing model. This is one half of bungalo 
semi (2 BR). Its composition is representative of timber-kit 
energy efficient public housing in the late 1990s in the 
North of Scotland.

Each room is explicitly represented as is the roof space 
and the crawl space. This allows transient occupancy patterns
to be studied as well as the impact on different infiltration
regimes.

This is an initial version of the model with the zone geometry
and composition specified. Room occupancy and environmental
controls have not yet been defined.  




Former-commit-id: c8f6b48414bba8199b6c46ef9891217414584ec2