Add a 3BR simi detached house which represents timber-kit construction which is in advance of late 1990s building standards in the far north of Scotland.
For flexibility, each room of the dwelling has been
represented as well as spaces which are likely to be
at variant temperatures (under stairs, crawl space
and roof void).

The building construction is assumed to be tighter
than normal in terms of infiltration.  

Model does not yet have operational characteristics
defined (for a later commit).




Former-commit-id: ea85de0f9bd7c7497b54196c2fc835dff336ec45