Addition of a small one BR bungalow. It has an attached entry and storage area.  Its composition is representative of timber-kit energy efficient public housing in the late 1990s in the North of Scotland.
Each room is explicitly represented as is the roof space 
and the crawl space. This allows transient occupancy patterns
to be studied as well as the impact on different infiltration
regimes.

This is an initial version of the model with the zone geometry
and composition specified. Room occupancy and environmental
controls have not yet been defined.  




Former-commit-id: 6c1d5397eb55699dbbb3ff8cd58e181d6fddd22d