Add a four level shallow plan office building which includes core zones at each level, entry, the ground and 1st level are low resolution, the 2nd level is subdivided by orientation and the upper level is again, lower resolution. Ceiling voids are treated explicitly as is the roof space.
There are two zones representing mixing boxes but they are
place holders for future variants which use flow networks to represent
a slightly temperated fresh air supply.

There is diversity included in weekday operations. Fresh air is
handled via scheduled infiltration in this model. Foundation construction
updated and matched to ground temperatures. There are no shanding devices
in the model but insolation calculations have been run. 

There are several scripts included in the model cfg folder. These
were created in an earlier project (IEA Annex 46 and as well as the
IMAGE project) and are not yet operational.
 



Former-commit-id: 240a145c32e406f73bf1cff9747b9eda42583489