- Update to the 2 BR wood insulated bungalow test model   to include control and the P1M occupancy pattern so that   it can be used in a similar way to the other test models.
Testing
- Interactive assessments of the average warm and cold comfort
  versions of the model. Looked at room temperatures.
  
Details of the model variants:

This is one half of bungalo semi (2 BR). Its composition is 
representative of timber energy efficient public housing 
in the late 1990s in the North of Scotland.

Each room is explicitly represented as is the roof space 
and the crawl space. This allows transient occupancy patterns
to be studied as well as the impact on different infiltration
regimes.

2BR_bungalow_p1m_avg.cfg - P1M occupancy with average comfort
2BR_bungalow_p1m_cool.cfg - P1M occupancy with cold comfort
2BR_bungalow_p1m_warm.cfg - P1M occupancy with warm comfort




Former-commit-id: b581a7b7ab2a261241ddf2d490c29624b762e39f