Basically re-enabled additional h3k output (mainly for CFC).
The output (including zone name) to "dictionary" looks as expected:
  "bui/*/sur_*/CFC_Uact_*","TheSpace*07*TGU","Actual CFC IGU U-value","(W/(m2 K))"
  "bui/*/sur*/CFC_Unom_*","TheSpace*07*TGU","Nominal ISO CFC IGU U-value","(W/(m2 K))"
  "bui/z_*/s_*/AmbRT","01*04","Ambient radiant temperature for exterior surface","(oC)"

Using (as examples)
  <step_variable>bui/TheSpace/sur_07/CFC_Uact_TGU</step_variable>
  <step_variable>bui/*/sur*/CFC_Unom_*</step_variable>
  <step_variable>bui/z_*/s_*/AmbRT</step_variable>

in input.xml leads to mangled variable names in out.csv and only output for the second "*"-version is generated:
  bui:*:sur*:**:*:sur*:CFC_Unom_* (W/(m2 K)),
  bui:z_***:s_*:AmbRT (oC),

The strings (the wild card tokens) are "doubled" and not expanded. The bug must be in whatever function builds the names for output. For some reason, this name building seems to work with the original long strings "zone_*" and "surface_*" (with two "*", only, however).

Compile and run on OS X O.K.

Results as such as expected.



Former-commit-id: 209061fd3835d8ffc0a40191bec905213d1fbfef