Extend CFC files with version number. Add output of (glass) layer names from GSLEdit and also IGDB index. Add report on this data in QA output.
There seems to be a problem on Cygwin and OS X in regard to (likely) stack memory. In any case, changes (addition of a few "mcom" sized arrays) led to inexplicable crashes until reducing the number of zones. Similar problems encountered some commits ago when increasing size for detailed 3d model - it would be very interesting to find a fix for this problem!

C, r & t on OS X seems O.K.

Tested with small model:
- used old format CFC files, O.K.
- re-imported GSLEdit file into this model, O.K.
- added new surface with CFC construction, O.K.



Former-commit-id: 68cb7f8f97af5e364e12e89feba3201da7323502