This commit addresses a logical error introduced onto development branch at r6485, and recently reported by Adam Wills. r6485 restructures management of infiltration and ventilation calculations in subsys.f, allowing models to exploit ESP-r's various infiltration and ventilation calculation methods simultaneously in a single simulation, and to switch between these using user-defined logic. As part of these modifications, I also enabled support for imposing scheduled ventilaiton flows via the opr file on top of AIM-2 computed infiltration flows.
Both the scheduled infiltration and ventilation flows defined in the opr file rely 
on the same calendar day common block variables to establish the current schedule 
entry. My initial implementation overlooked this fact. My code initialized the 
relevant day variables (ID1 and ID2) when scheduled infiltration was active, but 
not when the AIM-2 model was in use. The result was that scheduled ventilation could 
not be imposed simultaneously with AIM-2 computed infiltration. The code also 
attempted to read a 0-index-array location anytime the AIM-2 model was active, which 
caused a fatal bounds-check error in versions of bps compiled with the '--extra-debug' 
option. 

This change reorganizes code in subsys.F to define ID1 and ID2 regardless of whether 
the AIM-2 model is active. It has two functional effects upon the simulator: 

 - Versions of bps compuled with '--extra-debug' can now exercise the AIM-2 model 
   without crashing. 

 - All versions of bps can correctly impose AIM-2 computed infiltration and scheduled 
   ventilation simultaneously. This new capability affects a single model in the 
   test suite: plt_lookup_table. This is the only model that uses both aim-2 computed
   infiltration and scheduled ventilation---which ironically has no relevance of the 
   intent of plt_look_up_table. 

 




Former-commit-id: b193f32448402ef386f4ea7bf8cfcc8c13df6281