This commit reintroduces r7506 in preparation for submission to development_branch.
This commit corrects a long-standing bug in the non-HVAC electric 
loads power-only component, which imposes occupant-driven electric 
loads (as defined in a .fcl file) onto the electric network. The 
component contained a single additional input which could be used to 
scale the defined load as it was imposed on the network. The origninal 
author (me) carelessly cast this variable as a integer, which meant 
that any value less than one was intrepreted as zero (negating the 
electric loads entirely). 

This commit correctly casts the variable as a real, allowing floating 
inputs to be used when scaling the electric loads. 





Former-commit-id: 900e3325ae79fbe950f0a60e165bdbd384173e32