Add output of plant component additional data values to xml output.
plantc.db1.a sadly does not contain enough information to generate meaningful entries for all plant components that have additional data. Also, there does not seem to be a readily available common with the additional data stored. Res seems to build the description "on the fly" from the data base.

Added a simple "string filling" function to work around this limitation. Not nice, but seems to work.

Compile and run on OS X O.K., dictionary creation as expected.

ToDo: Check output values.

Former-commit-id: 9f9a34e53f1bfc474ebadad193f8a0f11fac841e