Introduced patches in the code so that if a tag *pausecas yes is added into a model configuration file in the line after a zone operation file has been defined then bps will pause with a user dialog at each time step asking if the zone operation file should be  rescanned. If answer no the casual.F logic proceeds as normal, if yes then in mzcasi the zone operation file is re-scanned and mzcasg is called to re-fill the casual gain arrays from which the future values are calculated.
Testing:

I took a small building model and setup three operation files
for an exhibition room.  A base operation file exhibition_24a.opr
which has 24 periods for occupants, lights & equipment on weekdays.  I copied
this into two other operation files which I called even and odd and edited
the even file casual gains.

I made a variant cfg file to point at the even operation file and another to
point to the odd operation file and ran assessments and generated trace.

I made another variant with the *pausecas token in cfg file and said no on the
startup day and then switched to even at 1h30 and back to odd at 3h00 and
then to even at 9h00.  The switch was done with a manual cp command in the
zones folder and  I ran the assessments at 30 minute time steps.

The trace for the mixed case shows the re-reading
of the operation file and the graphed values were updated
are as expected, but a closer look needed to confirm). 

I opened up res for the three difference results files and the mixed plots
out mixed as expected (a closer look needed to confirm).  The energy
balance for the zone I selected is balanced ;-). Thus the initial indicators
are that the manual copy and refresh is transparent to res because it is
just dumbly reading what was saved in the results file.

This approach allows models without the extra tag to proceed as normal. In
models which have the extra tag only those zones with the tag produce
requests for rescan.  This supports research into changes in operations
to reflect observations on site. Probably DO NOT commit to development_branch.




Former-commit-id: c71235e4b640503ca72ad0414ad5a5cf0d16c47a