- Add shell script and python scripts in src/bin to import   materials constructions and geometry from EnergyPlus IDF   files (see notes below) Python from Andrew Cowie of ESRU.
./convert_idf_2_espr.sh
Script to be run after import_geometryAndAttribution_fromEplusModel.py
and before associate_geometryAndDatabases_fromEplusModel.py. It
creates ESP-r folder structure as well as boilerplate cfg file. It
is passed the name of the IDF file and the root name of the ESP-r model.
first importing databases from the idf file

The method has been tested on a variety of EPP 8.5 models and the
 following caviats apply:
a) only the initial 30 characters of material & construction names used
b) zone & surface names must not include blanks
c) only material, construction and geometry tokens are parsed.
d) glazing optics are not parsed - you have to set these.
e) an ESP-r folder structure is created. You will have to re-name
   the mode cfg & cnn files, set location year etc.