Addition of Dockerfile
The Dockerfile holds the configuration/instructions for building a Docker image based on Ubuntu 18.04 that is set up to compile and run ESP-r. Instructions on how to use the Dockerfile have been added to the [wiki](https://github.com/ESP-rCommunity/ESP-rSource/wiki/Docker-Set-up). Currently instructions are provided for Windows users. Docker, however, runs natively on Linux and updating the wiki for those users is on the TODO list.